feat: landing page
customize time and timezone
