+++ src/app/desktop/desktop.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, Input } from '@angular/core';
 import * as data from '../speakers.json';
 
 @Component({
@@ -12,13 +12,10 @@ export class DesktopComponent implements OnInit {
   speakers: any[] = [];
   totalNumSpeaker: number = 0;
 
-  conferenceTime: string = '';
-  timezone: string = '';
-  startTimeUTC: number = 14;
-  endTimeUTC: number = 18;
-
   options = { threshold: 0.6 };
   
+  @Input() conferenceTime: string = '';
+  @Input() timezone: string = '';
 
   constructor() { }
 
@@ -26,7 +23,6 @@ export class DesktopComponent implements OnInit {
     this.speakers = this.speakerData.speakers;
     this.totalNumSpeaker = this.speakers.length;
     this.initializeAnimation();
-    this.customizeTime();
   }
   
   getImgSrc(name: string): string {
@@ -69,64 +65,5 @@ export class DesktopComponent implements OnInit {
     });
   }
 
-  customizeTime(): void {
-    // set up timezone
-    const offset = new Date().getTimezoneOffset();
-    if (offset > 0) {
-      this.timezone = timezoneMap1[offset];
-    } else {
-      this.timezone = timezoneMap2[-offset];
-    }
-    
-    // set up time
-    const hourDiff = offset / 60;
-    let startTag = '';
-    let endTag = '';
-    let localStart = this.startTimeUTC - hourDiff;
-    let localEnd = this.endTimeUTC - hourDiff;
-    if (localStart > 12) {
-      localStart = localStart % 12;
-      startTag = 'PM';
-    } else {
-      startTag = 'AM';
-    }
-    if (localEnd > 12) {
-      localEnd = localEnd % 12;
-      endTag = 'PM';
-    } else {
-      endTag = 'AM';
-    }
-    this.conferenceTime = localStart + ' ' + startTag + ' - ' + localEnd + ' ' + endTag;
-  }
-
 }
 
-const timezoneMap1: { [key: number]: string } = {
-  // GMT-7
-  420 : 'PDT',
-  // GMT-6
-  360 : 'CST',
-  // GMT-5
-  300 : 'EST',
-  // GMT-4
-  240 : 'EDT',
-  // GMT-3
-  180 : 'ADT',
-  // GMT-2
-  120 : 'GMT-2',
-  // GMT-1
-  60 : 'GMT-1'
-};  
-
-const timezoneMap2: { [key: number]: string } = {
-  // GMT
-  0 : 'WET',
-  // GMT+1
-  60 : 'CET',
-  // GMT+2
-  120 : 'EET',
-  // GMT+3
-  180 : 'EST',
-  // GMT+4
-  240 : 'MSK'
-};
