+++ src/app/mobile/mobile.component.html
@@ -41,8 +41,8 @@
             <div class="separator"></div>
                 <span class="label">TIME</span><br>
                 <div class="date-row">
-                    <span class="time-display">2 PM - 6 PM</span>
-                    <span class="superscript">CEST</span>
+                    <span class="time-display">{{ conferenceTime }}</span>
+                    <span class="superscript">{{ timezone }}</span>
                 </div>
             <div class="separator"></div>
                 <span class="label">ABOUT</span>
